Crate noodles_tabix

noodles-tabix handles the reading and writing of the tabix format.

A tabix (TBI) is an index file typically used to allow random access of an accompanied file that is

bgzipped,

tab-delimited,

grouped by reference sequence name, and

coordinate sorted by start position.

It can be used to find relevant records for a given genomic region.

Examples

Read a tabix file

use noodles_tabix as tabix;
let index = tabix::read("sample.vcf.gz.tbi")?;

Re-exports

pub use self::index::Index;

Modules

index

Tabix index and fields.

Structs

Reader

A tabix reader.

Writer

A tabix writer.

Functions

read

Reads the entire contents of a tabix index.

write

Writes a tabix index to a file.