Struct noodles_tabix::index::reference_sequence::ReferenceSequence

pub struct ReferenceSequence { /* private fields */ }

A tabix index reference sequence.

Implementations

impl ReferenceSequence

pub fn new(
bins: Vec<Bin>,
intervals: Vec<VirtualPosition>,
metadata: Option<Metadata>
) -> Self

Creates a tabix index reference sequence.

Examples

use noodles_tabix::index::ReferenceSequence;
let reference_sequence = ReferenceSequence::new(Vec::new(), Vec::new(), None);

pub fn bins(&self) -> &[Bin]

Returns the list of bins in the reference sequence.

This list does not include the metadata pseudo-bin (bin 37450). Use Self::metadata instead.

Examples

use noodles_tabix::index::ReferenceSequence;
let reference_sequence = ReferenceSequence::new(Vec::new(), Vec::new(), None);
assert!(reference_sequence.bins().is_empty());

pub fn intervals(&self) -> &[VirtualPosition]

Returns the list of 16 kbp intervals that make up the linear index.

Examples

use noodles_tabix::index::ReferenceSequence;
let reference_sequence = ReferenceSequence::new(Vec::new(), Vec::new(), None);
assert!(reference_sequence.intervals().is_empty());

pub fn query<I>(&self, interval: I) -> Result<Vec<&Bin>> where
I: Into<Interval>,

Returns a list of bins in this reference sequence that intersects the given range.

The interval values are 1-based.

Examples

use noodles_core::Position;
use noodles_tabix::index::ReferenceSequence;

let reference_sequence = ReferenceSequence::new(Vec::new(), Vec::new(), None);
let start = Position::try_from(8)?;
let end = Position::try_from(13)?;

let query_bins = reference_sequence.query(start..=end)?;
assert!(query_bins.is_empty());

pub fn min_offset(&self, start: Position) -> VirtualPosition

Finds in minimum start offset in the linear index for a given start position.

start is 1-based.

Examples

use noodles_bgzf as bgzf;
use noodles_core::Position;
use noodles_tabix::index::ReferenceSequence;

let reference_sequence = ReferenceSequence::new(Vec::new(), Vec::new(), None);
let start = Position::try_from(13)?;

assert_eq!(
    reference_sequence.min_offset(start),
    bgzf::VirtualPosition::from(0)
);